Use linux Miniconda installer
It looks like this job is now getting farmed out to `nott.stsci.edu`, which is a Linux machine, so we should use the linux Miniconda installer.

If possible, it would be nice to have some control over which type of machine actually runs this Jenkins job.  Will have to look into that.